+++ client-common/src/main/java/org/realityforge/replicant/client/runtime/AreaOfInterestServiceImpl.java
@@ -123,7 +123,7 @@ public class AreaOfInterestServiceImpl
   {
     if ( _subscriptions.containsKey( descriptor ) )
     {
-      throw new SubscriptionExistsException();
+      throw new SubscriptionExistsException( _subscriptions.get( descriptor ) );
     }
     else
     {