+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.5.90:
+* Add fields to the `SubscriptionExistsException`,`SubscriptionAlreadyRequiredException` and
+  `SubscriptionInactiveException` exceptions to aid debugging the application.
 * Fix `ChannelMetaData` so that it supports filtered type graphs.
 
 ## 0.5.89: